Fix pbkdf, pk padding and ECDH registration for static linking.
With this change the tests pass when linked against a static library
built in the normal (non-amalgamation) fashion.

Remove the restriction in configure.py, and have circleci build the
clang static build as a non-amalg.
