+++ src/lib/pbkdf/info.txt
@@ -7,7 +7,3 @@ base
 <header:public>
 pbkdf.h
 </header:public>
-
-<header:internal>
-pbkdf_utils.h
-</header:internal>
