+++ src/lib/pbkdf/pbkdf2/pbkdf2.cpp
@@ -5,15 +5,13 @@
 * Botan is released under the Simplified BSD License (see license.txt)
 */
 
-#include <botan/internal/pbkdf_utils.h>
 #include <botan/pbkdf2.h>
+#include <botan/lookup.h>
 #include <botan/get_byte.h>
 #include <botan/internal/xor_buf.h>
 #include <botan/internal/rounding.h>
 
 namespace Botan {
-
-BOTAN_REGISTER_NAMED_T(PBKDF, "PBKDF2", PKCS5_PBKDF2, PKCS5_PBKDF2::make);
 
 PKCS5_PBKDF2* PKCS5_PBKDF2::make(const Spec& spec)
    {
