+++ src/lib/pk_pad/eme_pkcs1/eme_pkcs.cpp
@@ -5,12 +5,9 @@
 * Botan is released under the Simplified BSD License (see license.txt)
 */
 
-#include <botan/internal/pad_utils.h>
 #include <botan/eme_pkcs.h>
 
 namespace Botan {
-
-BOTAN_REGISTER_EME_NAMED_NOARGS(EME_PKCS1v15, "PKCS1v15");
 
 /*
 * PKCS1 Pad Operation
