+++ src/lib/pk_pad/emsa1/emsa1.cpp
@@ -5,12 +5,9 @@
 * Botan is released under the Simplified BSD License (see license.txt)
 */
 
-#include <botan/internal/pad_utils.h>
 #include <botan/emsa1.h>
 
 namespace Botan {
-
-BOTAN_REGISTER_EMSA_1HASH(EMSA1, "EMSA1");
 
 namespace {
