+++ src/lib/pk_pad/emsa_pkcs1/emsa_pkcs1.cpp
@@ -11,9 +11,7 @@
 
 namespace Botan {
 
-namespace {
-
-EMSA* make_pkcs1v15(const EMSA::Spec& spec)
+EMSA* EMSA_PKCS1v15::make(const EMSA::Spec& spec)
    {
    if(spec.arg(0) == "Raw")
       return new EMSA_PKCS1v15_Raw;
@@ -24,10 +22,6 @@ EMSA* make_pkcs1v15(const EMSA::Spec& spec)
       }
    return nullptr;
    }
-
-}
-
-BOTAN_REGISTER_NAMED_T(EMSA, "EMSA_PKCS1", EMSA_PCS1v15, make_pkcs1v15);
 
 namespace {
