+++ src/lib/pk_pad/emsa_raw/emsa_raw.cpp
@@ -10,8 +10,6 @@
 
 namespace Botan {
 
-BOTAN_REGISTER_EMSA_NAMED_NOARGS(EMSA_Raw, "Raw");
-
 /*
 * EMSA-Raw Encode Operation
 */
