+++ src/lib/pubkey/ecdh/ecdh.cpp
@@ -12,6 +12,8 @@
 
 namespace Botan {
 
+ECDH_PublicKey::ECDH_PublicKey() {}
+
 namespace {
 
 /**
