+++ src/scripts/ci/circle/clang-static-debug.sh
@@ -5,6 +5,6 @@ which shellcheck > /dev/null && shellcheck "$0" # Run shellcheck on this if avai
 BUILD_NICKNAME=$(basename "$0" .sh)
 BUILD_DIR="./build-$BUILD_NICKNAME"
 
-./configure.py --with-build-dir="$BUILD_DIR" --build-mode=debug --cc=clang --disable-shared --via-amalgamation
+./configure.py --with-build-dir="$BUILD_DIR" --build-mode=debug --cc=clang --disable-shared
 make -j 2 -f "$BUILD_DIR"/Makefile
 "$BUILD_DIR"/botan-test
